+++ index.js
@@ -1,7 +1,9 @@
-const CrudBaseService = require('./lib/base-service');
+const CrudBaseService = require('./lib/crud-base-service');
+const CrudBaseServiceFull = require('./lib/crud-base-service-full');
 const TestUtils = require('./lib/test-utils');
 
 module.exports = {
   CrudBaseService,
+  CrudBaseServiceFull,
   TestUtils,
 };
