+++ index.spec.js
@@ -1,10 +1,15 @@
 const condorMongoose = require('./index');
-const BaseService = require('./lib/base-service');
+const CrudBaseService = require('./lib/crud-base-service');
+const CrudBaseServiceFull = require('./lib/crud-base-service-full');
 const TestUtils = require('./lib/test-utils');
 
 describe('condor-mongoose', () => {
   it('should expose CrudBaseService', () => {
-    expect(condorMongoose.CrudBaseService).toBe(BaseService);
+    expect(condorMongoose.CrudBaseService).toBe(CrudBaseService);
+  });
+
+  it('should expose CrudBaseServiceFull', () => {
+    expect(condorMongoose.CrudBaseServiceFull).toBe(CrudBaseServiceFull);
   });
 
   it('should expose TestUtils', () => {
