+++ lib/crud-base-service-full.integration.spec.js
@@ -7,16 +7,16 @@ const proxyquire = require('proxyquire');
 const _ = require('lodash');
 
 const CrudBaseService = require('./crud-base-service');
-const BaseService = require('./base-service');
+const CrudBaseServiceFull = require('./crud-base-service-full');
 const Model = require('./../spec/stubs/models/model/model');
 const RelatedModel = require('./../spec/stubs/models/related-model/related-model');
 const Service = require('./../spec/stubs/services/service');
 
-const TestUtils = proxyquire('./test-utils', {'BaseService': BaseService});
+const TestUtils = proxyquire('./test-utils', {'BaseService': CrudBaseServiceFull});
 
 mongoose.Promise = Promise;
 
-describe('Base Service', () => {
+describe('Crud Base Service Full', () => {
   let condor, client, model1, relatedModel1, relatedModel2, child;
   const protoPath = './spec/stubs/proto/sample.proto';
   const invalidId = '5894e68edfe69eb5159c5665';
@@ -40,18 +40,18 @@ describe('Base Service', () => {
     const newRelatedModel = new RelatedModel(defaultRelatedModel1);
     return newRelatedModel.save()
     .then((relatedModel) => {
-      relatedModel1 = BaseService.getModelObject(relatedModel);
+      relatedModel1 = CrudBaseServiceFull.getModelObject(relatedModel);
       const defaultRelatedModel2 = {'name': 'defaultRelatedModel2'};
       const newRelatedModel = new RelatedModel(defaultRelatedModel2);
       return newRelatedModel.save();
     }).then((relatedModel) => {
-      relatedModel2 = BaseService.getModelObject(relatedModel);
+      relatedModel2 = CrudBaseServiceFull.getModelObject(relatedModel);
       const defaultModel1 = {'name': 'Juan Pablo', 'age': 33, 'married': false,
         'children': [child], 'child': child, 'relatedModels': [relatedModel1.id]};
       const newModel = new Model(defaultModel1);
       return newModel.save();
     }).then((model) => {
-      model1 = BaseService.getModelObject(model);
+      model1 = CrudBaseServiceFull.getModelObject(model);
     }).then(done);
   });
 
@@ -76,13 +76,13 @@ describe('Base Service', () => {
   });
 
   it('should extend from CrudBaseService', () => {
-    expect(BaseService.prototype instanceof CrudBaseService).toBeTruthy();
+    expect(CrudBaseServiceFull.prototype instanceof CrudBaseService).toBeTruthy();
   });
 
   describe('constructor()', () => {
     it('should create instances of BaseService', () => {
-      const baseService = new BaseService(Model);
-      expect(baseService instanceof BaseService).toBeTruthy();
+      const crudBaseServiceFull = new CrudBaseServiceFull(Model);
+      expect(crudBaseServiceFull instanceof CrudBaseServiceFull).toBeTruthy();
     });
 
     describe('when mongoose is not connected', () => {
@@ -92,7 +92,7 @@ describe('Base Service', () => {
 
       it('should throw an error', () => {
         expect(() => {
-          new BaseService(); // eslint-disable-line
+          new CrudBaseServiceFull(); // eslint-disable-line
         }).toThrowError('mongoose is not connected');
       });
     });
