+++ lib/crud-base-service-full.js
@@ -6,7 +6,7 @@ const ActionEnum = Object.freeze(
     {'PUSH': 0, 'ADD_TO_SET': 1, 'REMOVE': 2, 'REPLACE': 3, 'UPDATE': 4}
   );
 
-class BaseService extends CrudBaseService {
+class CrudBaseServiceFull extends CrudBaseService {
   constructor(Model) {
     if (mongoose.connection.readyState !== 1) {
       throw new Error('mongoose is not connected');
@@ -39,7 +39,7 @@ function createAdditionalMethods(Model) {
 
 function addMethods(methods, field) {
   Object.keys(methods).forEach((key) => {
-    BaseService.prototype[key + capitalizeFirstLetter(field)] = methods[key];
+    CrudBaseServiceFull.prototype[key + capitalizeFirstLetter(field)] = methods[key];
   });
 }
 
@@ -97,7 +97,7 @@ function runRelatedModelRequest(Model, request, action) {
 }
 
 function executeSubDocumentAction(request, model, action) {
-  const fieldName = BaseService.getSubDocumentFieldName(request);
+  const fieldName = CrudBaseServiceFull.getSubDocumentFieldName(request);
   if (action === ActionEnum.REPLACE) {
     model[fieldName] = [];
   }
@@ -127,7 +127,7 @@ function executeSubDocumentAction(request, model, action) {
 }
 
 function executeRelatedModelAction(request, model, action) {
-  const fieldName = BaseService.getSubDocumentFieldName(request);
+  const fieldName = CrudBaseServiceFull.getSubDocumentFieldName(request);
   if (action === ActionEnum.REPLACE) {
     model[fieldName] = [];
   }
@@ -183,4 +183,4 @@ function capitalizeFirstLetter(string) {
   return string[0].toUpperCase() + string.slice(1);
 }
 
-module.exports = BaseService;
+module.exports = CrudBaseServiceFull;
