+++ lib/test-utils.js
@@ -1,7 +1,7 @@
 const mockgoose = require('mockgoose');
 const mongoose = require('mongoose');
 const Promise = require('bluebird');
-const BaseService = require('./base-service');
+const BaseService = require('./crud-base-service-full');
 
 mongoose.Promise = Promise;
