+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "condor-mongoose",
-  "version": "0.0.5",
+  "version": "0.1.0",
   "description": "Utils to work with mongoose and condor GRPC framework",
   "main": "index.js",
   "scripts": {
