+++ spec/stubs/services/service.js
@@ -1,4 +1,4 @@
-const BaseService = require('./../../../lib/base-service');
+const BaseService = require('./../../../lib/crud-base-service-full');
 const Model = require('../models/model/model');
 module.exports = class extends BaseService {
   constructor() {
